+++ src/TRegx/CleanRegex/Match/Details/ReplaceMatchImpl.php
@@ -139,9 +139,19 @@ class ReplaceMatchImpl implements ReplaceMatch
         return $this->match->offset();
     }
 
+    public function tail(): int
+    {
+        return $this->match->tail();
+    }
+
     public function byteOffset(): int
     {
         return $this->match->byteOffset();
+    }
+
+    public function byteTail(): int
+    {
+        return $this->match->byteTail();
     }
 
     public function setUserData($userData): void